Changes to the datatype

+++ nth_fibonacci/src/main.rs
@@ -8,17 +8,17 @@ fn main() {
         .read_line(&mut index)
         .expect("Failed to read the index value!");
     
-    let index:u32 = index.trim().parse().expect("Invalid index!");
+    let index:u64 = index.trim().parse().expect("Invalid index!");
 
-    let fibonacci_number: u32 = 
+    let fibonacci_number: u64 = 
     if (index == 1) || (index == 2) {index-1}
     else {
-        let mut prev: u32 = 0;
-        let mut next: u32 = 1;
-        let mut temp_index: u32 = index;
+        let mut prev: u64 = 0;
+        let mut next: u64 = 1;
+        let mut temp_index: u64 = index;
 
         while temp_index-2 != 0 {
-            let temp: u32 = next;
+            let temp: u64 = next;
             next += prev;
             prev = temp;
